# make decirc function public