--- a/index.d.ts
+++ b/index.d.ts
@@ -3,6 +3,7 @@ declare function stringify(data: any): string;
 declare namespace stringify {
   export function stable(data: any): string;
   export function stableStringify(data: any): string;
+  export function decirc(val: any, k: string, stack: any[], parent?: any): void;
 }
 
 export default stringify;