--- a/test.js
+++ b/test.js
@@ -3,58 +3,6 @@ const fss = require('./')
 const clone = require('clone')
 const s = JSON.stringify
 
-// this custom replacer will add a json pointer $ref pointing to the referenced js object
-test('decycle supports custom replacer', function (assert) {
-  const fixture = {
-    name: 'Tywin Lannister',
-    house2: {
-      inner: {}
-    },
-    house: {
-      name: 'Lannister'
-    }
-  }
-
-  fixture.house2.inner = fixture.house
-  fixture.house.circle = fixture.house2.inner
-
-  const expected = {
-    name: 'Tywin Lannister',
-    house2: {
-      inner: {
-        name: 'Lannister',
-        circle: {
-          key: 'circle',
-          stackSize: 3,
-          parentName: 'Lannister',
-          $ref: '#/house2/inner'
-        }
-      }
-    },
-    house: {
-      name: 'Lannister',
-      circle: {
-        key: 'circle',
-        stackSize: 3,
-        parentName: 'Lannister',
-        $ref: '#/house2/inner'
-      }
-    }
-  }
-
-  const actual = fss.decycle(fixture, (val, k, stack, parent) => {
-    return {
-      key: k,
-      stackSize: stack.length,
-      parentName: parent.name,
-      $ref: `#${stack.map((s) => s[1]).join('/')}`
-    }
-  })
-
-  assert.deepEqual(actual, expected)
-  assert.end()
-})
-
 test('circular reference to root', function (assert) {
   const fixture = { name: 'Tywin Lannister' }
   fixture.circle = fixture
@@ -288,5 +236,80 @@ test('nested child circular reference in toJSON', function (assert) {
   })
   const actual = fss(o)
   assert.is(actual, expected)
+  assert.end()
+})
+
+// this custom replacer will add a json pointer $ref pointing to the referenced js object
+test('decycle supports custom replacer', function (assert) {
+  const fixture = {
+    definitions: {
+      Customer: {
+        properties: {
+          partners: {
+            items: {}
+          }
+        }
+      },
+      Partner: {
+        properties: {
+          customers: {
+            items: {}
+          }
+        }
+      }
+    }
+  }
+
+  fixture.definitions.Customer.properties.partners.items = fixture.definitions.Partner
+  fixture.definitions.Partner.properties.customers.items = fixture.definitions.Customer
+
+  const expected = {
+    definitions: {
+      Customer: {
+        properties: {
+          partners: {
+            items: {
+              properties: {
+                customers: {
+                  items: {
+                    $ref: '#/definitions/Customer'
+                  }
+                }
+              }
+            }
+          }
+        }
+      },
+      Partner: {
+        properties: {
+          customers: {
+            items: {
+              $ref: '#/definitions/Customer'
+            }
+          }
+        }
+      }
+    }
+  }
+
+  const actual = fss.decycle(fixture, (val, k, stack, parent) => {
+    let $ref = '#'
+
+    for (let i = 0; i < stack.length; i++) {
+      if (stack[i][1]) {
+        $ref += `/${stack[i][1]}`
+      }
+
+      if (stack[i][0] === val) {
+        break
+      }
+    }
+
+    return {
+      $ref
+    }
+  })
+
+  assert.deepEqual(actual, expected)
   assert.end()
 })