We find it quite useful to remove circular references without having to go through a full stringify and then re-parse.