@davidmarkclements 2018-04-11
If we expose it publically it also needs documentation and unit tests 

We may also want to consider the API