@marbemac 2018-04-11
Makes sense, I can do that - in general is this a change that you'd be interested in? For the simple case of needing to decycle an existing JS object, small objects jump from 100k -> 600k ops in our benchmarks, and large objects jump even more, so this has been great for us (rather than doing a `JSON.parse(safeStringify(obj))`).

Do you already have thoughts on possible changes to the API for this function?