@davidmarkclements 2018-04-12
Does this mean that there are potential cases where decirc would be used but stringify wouldn't? 

If that is true, decirc should be a separate module - and fast safe stringify can rely on it