@mcollina 2018-04-12
Definitely +1 in splitting. Add me as well (also on npm)!